fix(lab): история прогонов не падала на qty="1.0"; ресайз границ фреймов мышкой
- /api/v1/lab/campaigns кастил params->>'qty'/'avg_max' к ::int, а optimizer
  пишет их как float-строки ("1.0") -> 500, «История прогонов» пустела при
  «перебор 100%». Каст через ::numeric (тот же класс, что фильтр вырожденных).
- Backtest Lab: перетаскиваемый вертикальный сплиттер между конфигом и
  результатами; ширина панели запоминается в localStorage (btl.configW).

+++ trader/api/app.py
@@ -2228,8 +2228,8 @@ def create_app() -> FastAPI:
                    date_from, date_to
             FROM optimization_leaderboard
             WHERE campaign_run IS NOT NULL AND net_profit IS NOT NULL
-              AND COALESCE(NULLIF(params->>'qty','')::int, 1) <= 10
-              AND COALESCE(NULLIF(params->>'avg_max','')::int, 1) <= 10
+              AND COALESCE(NULLIF(params->>'qty','')::numeric, 1) <= 10
+              AND COALESCE(NULLIF(params->>'avg_max','')::numeric, 1) <= 10
             ORDER BY campaign_run, net_profit DESC NULLS LAST
             """)
         lead = {r["campaign_run"]: r for r in leaders}
